PR: Fixed File Path Issues (tested on Windows)
This pull request fixes an issue that occurred on Windows when using "Save and generate" in the "SteamVR Input" pane (in Unity). All default bindings were removed because the binding files were not found due to the script looking for just the file name (instead of using the full path). Currently, a warning is logged when GetBindingFileObject(path) / CheckPath(path) is called with an incomplete path but the path is automatically corrected (see method CheckPath(string path) in SteamVR_Input_ActionManifest_Manager. That warning could probably be removed.

This pull request also fixes a build-time issue that occurred when actions.json was not in the project root folder and its folder had not been previously created in the build target folder. This folder is now automatically created if it not exists.